backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for an AI-powered chatbot, capable of providing well-founded answers about the Unicamp 2026 admission exam"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "openai>=2.6.1",
    "pydantic-settings>=2.11.0",
    "sqlmodel>=0.0.27",
    "psycopg2-binary>=2.9.9",
    "authlib>=1.6.5",
    "itsdangerous>=2.2.0",
    "python-jose>=3.5.0",
    "alembic>=1.17.1",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.3",
    "langchain>=1.0.3",
    "langchain-text-splitters>=1.0.0",
    "bs4>=0.0.2",
    "langchain-openai>=1.0.2",
    "langchain-chroma>=1.0.0",
    "onnxruntime>=1.23.2",
    "asyncpg>=0.30.0",
    "pypdf>=4.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
    "pytest>=8.4.2",
    "types-authlib>=1.6.5.20251005",
    "types-python-jose>=3.5.0.20250531",
    "ty>=0.0.1a25",
]

[tool.ruff]
exclude = [
    "migrations/versions/**",
]

[tool.ty.src]
exclude = [
    "migrations/versions/**",
]
